// apps/admin/src/generated/api/schemas/updateUploadSettingsDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUploadSettingsDto {
  /** @minimum 1 */
  maxSizeMb: number;
  /** @minItems 1 */
  allowedMimeTypes: string[];
}